+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 
-## UNRELEASED
+## 1.0.0 - 2016-06-25
 
 ### Changed